requirements.txt
streamlit==1.28.0
psycopg2-binary==2.9.7
pandas==2.0.3
plotly==5.15.0
streamlit-authenticator==0.2.2
python-dotenv==1.0.0
qrcode[pil]==7.4.2
reportlab==4.0.4
Pillow==10.0.1
openpyxl==3.1.2